Brokerages/Fidelity/requirements.txt
fidelity-api>=0.0.16
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
